requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jinja2
python-dotenv
apscheduler
yfinance
pandas
requests
email-validator
